BackEnd/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.2.1
nltk==3.8.1
google-generativeai==0.3.2
requests==2.31.0
mysqlclient==2.2.4
python-dotenv==1.0.1 